# docs/api/swagger.yaml
definitions:
  models.APIError:
    properties:
      error:
        description: This field determines the raised error
        type: string
      message:
        description: This field determines the detailed information about raise error
        type: string
    type: object
  models.AccessLevelPermsSwagger:
    properties:
      access_level_perms_id:
        example: ed83a2ba-c55c-4297-b2ac-df7b02abdd7a
        type: string
      permission_set_id:
        type: string
    type: object
  models.Organization:
    type: object
  models.OrganizationDetailsResponse:
    properties:
      address:
        type: string
      email:
        type: string
      mobile:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  models.OrganizationOwnerResponse:
    properties:
      address:
        type: string
      email:
        type: string
      legal_national_id:
        type: string
      mobile:
        type: string
      name:
        type: string
      phone:
        type: string
      type:
        type: string
    type: object
  models.OrganizationResponse:
    properties:
      allow_negative_balance:
        type: boolean
      balance:
        type: number
      negative_balance_threshold:
        type: number
      organization_details:
        $ref: '#/definitions/models.OrganizationDetailsResponse'
      organization_id:
        example: ed83a2ba-c55c-4297-b2ac-df7b02abdd7a
        type: string
      organization_owner:
        $ref: '#/definitions/models.OrganizationOwnerResponse'
    type: object
  models.OrganizationShortInfo:
    properties:
      organization_id:
        description: This field determines the unique id of the organization. The
          id is in uuid v4 format
        example: ed83a2ba-c55c-4297-b2ac-df7b02abdd7a
        type: string
      organization_name:
        description: This field determines the name of the organization
        example: sample organization
        type: string
    type: object
  models.OrganizationalLevelPermsSwagger:
    properties:
      organizational_level_perms_id:
        example: ed83a2ba-c55c-4297-b2ac-df7b02abdd7a
        type: string
      permission_set_id:
        example: ed83a2ba-c55c-4297-b2ac-df7b02abdd7a
        type: string
    type: object
  models.PaymentLevelPermsSwagger:
    properties:
      payment_level_perms_id:
        example: ed83a2ba-c55c-4297-b2ac-df7b02abdd7a
        type: string
      permission_set_id:
        type: string
    type: object
  models.PermissionSetsSwagger:
    properties:
      access_level_perms:
        $ref: '#/definitions/models.AccessLevelPermsSwagger'
      organizational_level_perms:
        $ref: '#/definitions/models.OrganizationalLevelPermsSwagger'
      payment_level_perms:
        $ref: '#/definitions/models.PaymentLevelPermsSwagger'
      permission_sets_id:
        example: ed83a2ba-c55c-4297-b2ac-df7b02abdd7a
        type: string
      report_level_perms:
        $ref: '#/definitions/models.ReportLevelPermsSwagger'
      subscriber_group_id:
        example: ed83a2ba-c55c-4297-b2ac-df7b02abdd7b
        type: string
      subscriber_level_perms:
        $ref: '#/definitions/models.SubscriberLevelPermsSwagger'
    type: object
  models.ReportLevelPermsSwagger:
    properties:
      permission_set_id:
        type: string
      report_level_perms_id:
        example: ed83a2ba-c55c-4297-b2ac-df7b02abdd7a
        type: string
    type: object
  models.SubscriberGroupCreateResponse:
    properties:
      id:
        type: string
      message:
        type: string
      name:
        type: string
    type: object
  models.SubscriberGroupSwagger:
    properties:
      organization_id:
        type: string
      subscriber_group_description:
        type: string
      subscriber_group_id:
        type: string
      subscriber_group_name:
        type: string
      subscriber_group_permissions:
        $ref: '#/definitions/models.PermissionSetsSwagger'
    type: object
  models.SubscriberLevelPermsSwagger:
    properties:
      permission_set_id:
        type: string
      subscriber_level_perms_id:
        example: ed83a2ba-c55c-4297-b2ac-df7b02abdd7a
        type: string
    type: object
info:
  contact:
    email: ma.ahmadi1989@gmail.com
    name: Mahmoud Ahmadi
  description: This document covers the API endpoints exposed by Owl MNS Subscriber
    Profile Manager.
  title: Owl MNS - OSPM - API Reference
  version: "1.0"
paths:
  /organization:
    delete:
      consumes:
      - application/json
      description: \
      parameters:
      - description: Organization ID
        in: query
        name: id
        type: string
      - description: Organization Name
        in: query
        name: name
        type: string
      - description: 'Deletion Mode: hard/soft'
        in: query
        name: mode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organization successfully deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Delete an organization
      tags:
      - Organization
    get:
      description: \
      parameters:
      - description: includes soft deleted organizations (Optional)
        in: query
        name: list_all
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful Response
          schema:
            items:
              $ref: '#/definitions/models.OrganizationShortInfo'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: List all organizations
      tags:
      - Organization
    post:
      consumes:
      - application/json
      description: Adds a new organization to the system. The request body must contain
        the organization details.
      parameters:
      - description: Organization details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Organization'
      produces:
      - application/json
      responses:
        "201":
          description: Organization successfully added
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Add a new organization
      tags:
      - Organization
  /organization/{organization-id}/subscriber-group:
    get:
      consumes:
      - application/json
      description: Returns a list of all subscriber groups within an organization
      parameters:
      - description: Organization ID
        in: path
        name: organization-id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            items:
              $ref: '#/definitions/models.SubscriberGroupSwagger'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: List all Subscriber Groups
      tags:
      - Organization
  /organization/{organization-id}/subscriber-group/:
    post:
      consumes:
      - application/json
      description: Adds a new subscriber group within an organization
      parameters:
      - description: Subscriber Group ID
        in: path
        name: organization-id
        required: true
        type: integer
      - description: Subscriber Group Details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SubscriberGroupSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully added new subscriber group
          schema:
            $ref: '#/definitions/models.SubscriberGroupCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Add New Subscriber Group
      tags:
      - Organization
  /organization/{organization-id}/subscriber-group/{subscriber-group-id}:
    delete:
      consumes:
      - application/json
      description: Deletes a specific subscriber group within an organization by its
        ID
      parameters:
      - description: Subscriber Group ID
        in: path
        name: subscriber-group-id
        required: true
        type: integer
      - description: Subscriber Group ID
        in: path
        name: organization-id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Delete a Subscriber Group
      tags:
      - Organization
    get:
      consumes:
      - application/json
      description: Retrieves the details of a specific subscriber group within an
        organization
      parameters:
      - description: Organization ID
        in: path
        name: organization-id
        required: true
        type: integer
      - description: Subscriber Group ID
        in: path
        name: subscriber-group-id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/models.SubscriberGroupSwagger'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get Subscriber Group Detail
      tags:
      - Organization
    patch:
      consumes:
      - application/json
      description: Updates the settings of a specific subscriber group within an organization
        by its ID
      parameters:
      - description: Subscriber Group ID
        in: path
        name: subscriber-group-id
        required: true
        type: integer
      - description: Subscriber Group ID
        in: path
        name: organization-id
        required: true
        type: integer
      - description: Subscriber Group Settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SubscriberGroupSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Update Subscriber Group
      tags:
      - Organization
  /organization/recover/profile:
    patch:
      description: \
      parameters:
      - description: Organization ID
        in: query
        name: id
        type: string
      - description: Organization Name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organization successfully deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Recover s soft deleted organization
      tags:
      - Organization
  /organizations/profile:
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific organization identified
        by its name or ID.
      parameters:
      - description: Organization Name
        in: query
        name: name
        type: string
      - description: Organization ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful Response
          schema:
            $ref: '#/definitions/models.OrganizationResponse'
        "404":
          description: Organization Not Found
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Get organization profile by name or ID
      tags:
      - Organization
swagger: "2.0"